browserify - move core/createConfigSpy -> browserify/createModuleInspectorSpy
